requirements.txt
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
pandas>=2.0.0
openpyxl>=3.1.0

